NewObject must increment reference count when reusing an object
The object returned by NewObject must always be able to have
its reference count decremented by the caller. Previously
cached objects were not have a reference increment and so
the caller would have to know whether the object was cached
or created by the routine.

This caused negative reference counts in the tests.