Add Channel class
The python interface for the Channel class constructor has no source or
sink functions and so basic Channels reads from standard input (or the
SourceFile file) and write to standard output (or the SinkFile file).
However, source and sink functions can be supplied by extending the
basic Channel class and providing implementations for source and sink
methods in the subclass.