Clear the Card attribute automatically before reading from a FitsChan
This is almost always what is needed, but just in case people want to
read from somewhere in the middle of a FitsChan (e.g. if reading multiple
objects using native encoding), the read method now has an optional
argument that can be set False to prevent the FitsCHan being rewound.