Build without using configure
It turns out that there aren't many tests that configure
needs to perform.

This patch tests for strerror_r and strtok_r but does not worry
about HAVE_LONG_DOUBLE, HAVE_STDARG_H, HAVE_BACKTRACE, HAVE_EXECINFO_H
or HAVE_STAR_MEM_H.

The two that may cause backwards compatibility problems are the isnan
test and HAVE_NONANSI_SSCANF. For the sscanf issue it might be best to
have an explicit test in the test suite that would tickle the problem
(if we don't already have one) and then just put in a list of
operating systems that have a broken sscanf.