Handle PyArray_LONG vs PyArray_INT anomaly
On 64-bit the int32 data type turns up as a PyArray_INT type in C
but on 32-bit int32 becomes a PyArray_LONG. astRebinSeq was failing
a test in test_Mapping because of this on 32-bit systems.

This patch is a bit of a hack because it tests the size of long
and int and decides that LONG can be used safely if they are the
same size. There is probably a better approach.