Add UnitNormMap
Add new mapping UnitNormMap which takes a center
and maps a vector of the same dimensions
to a unit vector and norm relative to that center.
Thus the forward transform adds one dimension.
Add a unit test for the new mapping.
Add check... methods to test.py that allow testing various
aspects of mappings and other objects, such as persistence.